+++ src/export-const.js
@@ -1,0 +1,11 @@
+const FILTERS_AREA = document.querySelector(`.main__filter`);
+const CARDS_AREA = document.querySelector(`.board__tasks`);
+const WEEKDAYS = [`mo`, `tu`, `we`, `th`, `fr`, `sa`, `su`];
+const COLORLIST = [`black`, `yellow`, `blue`, `green`, `pink`];
+
+export default {
+  FILTERS_AREA,
+  CARDS_AREA,
+  WEEKDAYS,
+  COLORLIST
+};
